@@ -8,37 +8,39 @@
     "@material-ui/lab": "^4.0.0-alpha.58",
     "@material-ui/system": "^4.11.3",
     "add": "^2.0.6",
-    "axios": "^0.21.2",
-    "babel-jest": "^24.8.0",
+    "axios": "^1.1.3",
+    "babel-jest": "^29.2.0",
     "babel-plugin-named-asset-import": "^0.3.7",
-    "babel-preset-react-app": "^9.1.2",
-    "camelcase": "^5.2.0",
-    "case-sensitive-paths-webpack-plugin": "2.2.0",
+    "babel-preset-react-app": "^10.0.1",
+    "camelcase": "^7.0.0",
+    "case-sensitive-paths-webpack-plugin": "2.4.0",
     "cors": "^2.8.5",
-    "css-loader": "2.1.1",
-    "dotenv": "^6.2.0",
-    "dotenv-expand": "4.2.0",
-    "eslint-config-react-app": "^5.2.1",
-    "eslint-loader": "2.2.1",
-    "eslint-plugin-flowtype": "3.13.0",
-    "eslint-plugin-import": "2.18.2",
-    "eslint-plugin-jsx-a11y": "6.2.3",
-    "eslint-plugin-react": "7.14.3",
-    "eslint-plugin-react-hooks": "^1.6.1",
-    "eslint-utils": "^1.4.3",
+    "css-loader": "6.7.1",
+    "dotenv": "^16.0.3",
+    "dotenv-expand": "9.0.0",
+    "eslint-config-react-app": "^7.0.1",
+    "eslint-loader": "4.0.2",
+    "eslint-plugin-flowtype": "8.0.3",
+    "eslint-plugin-import": "2.26.0",
+    "eslint-plugin-jsx-a11y": "6.6.1",
+    "eslint-plugin-react": "7.31.10",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-utils": "^3.0.0",
     "express": "^4.17.1",
     "express-http-proxy": "^1.6.2",
+    "history": "^5.3.0",
+    "joi": "^17.6.3",
     "morgan": "^1.10.0",
     "node-fetch": "^3.1.1",
-    "nodemon": "^1.19.4",
-    "nuka-carousel": "^4.7.7",
-    "react": "^16.14.0",
-    "react-device-detect": "^1.17.0",
-    "react-dom": "^16.14.0",
-    "react-global-hook": "^1.4.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "^3.4.4",
-    "use-global-hook": "^0.1.13",
+    "nodemon": "^2.0.20",
+    "nuka-carousel": "^5.4.0",
+    "react": "^18.2.0",
+    "react-device-detect": "^2.2.2",
+    "react-dom": "^18.2.0",
+    "react-global-hook": "^2.0.0",
+    "react-router-dom": "^6.4.2",
+    "react-scripts": "5.0.1",
+    "use-global-hook": "^0.3.0",
     "yarn": "^1.22.10"
   },
   "scripts": {
@@ -66,6 +68,6 @@
   "main": "index.js",
   "license": "MIT",
   "devDependencies": {
-    "concurrently": "^4.1.2"
+    "concurrently": "^7.4.0"
   }
 }
